Move table view change methods to a UITableViewController extension
